Задача "Магазин"
Рефакторинг:
Удалил ненужные методы и классы, перенёс методы из DictionaryExtension в DictionaryExtensions. Перенёс часть классов в папку Entities
