+++ Shop/Entities/Storage.cs
