--- Shop/ProductName.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Shop
-{
-    public static class ProductName
-    {
-        public static Dictionary<Type, string> ProductNames = new Dictionary<Type, string>()
-        {
-            {typeof(Apple), "Яблоко"},
-            {typeof(Peach), "Грушка"},
-            {typeof(Candy), "Конфеты"},
-            {typeof(Backpack), "Рюкзак"}
-        };
-    }
-}
