--- Shop/Structure.txt
@@ -1,15 +1,0 @@
-﻿1) Покупатель:
-+Имеет: Деньги, вместимость сумки, вещи в сумке
-Может: Покупать, Смотреть вещи в своей сумке
-
-2) Хранилище:
-Имеет: Товары, вместимость хранилища
-Может: Выдавать товары из себя покупателю, показывать товары
-
-3) Продавец:
-Имеет: Деньги в кассе, список товаров
-Может: Продавать, Показывать товары
-
-4) Товар:
-+Имеет: Цена, вес
-Может: +Создаваться, изменять ценник
